# api/docs/swagger.yaml
definitions:
  controller.createSandboxBody:
    properties:
      code:
        type: string
      language:
        type: string
      name:
        type: string
    type: object
  controller.signInBody:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  controller.signUpBody:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  controller.updateSandboxBody:
    properties:
      code:
        type: string
      language:
        type: string
      name:
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  description: YARCEE API swagger documentation
  title: YARCEE API
  version: "1.0"
paths:
  /api/sandbox/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Delete sandbox
      tags:
      - Sandbox
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Get user sandboxes
      tags:
      - Sandbox
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.createSandboxBody'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Create sandbox
      tags:
      - Sandbox
  /api/sandbox/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Get sandbox
      tags:
      - Sandbox
    put:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.updateSandboxBody'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Update sandbox
      tags:
      - Sandbox
  /api/sign-in/:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.signInBody'
      produces:
      - application/json
      responses: {}
      summary: Sign in
      tags:
      - Auth
  /api/sign-up/:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.signUpBody'
      produces:
      - application/json
      responses: {}
      summary: Sign up
      tags:
      - Auth
  /api/user/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: User auth details
      tags:
      - User
  /health_check/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      summary: Health check
      tags:
      - Health check
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"